+++ tests/manage/test_profile_access.py
@@ -17,8 +17,8 @@ from subprocess import PIPE, Popen
 import psutil
 import pytest
 
-from aiida.backends.managers.profile_access import ProfileAccessManager
 from aiida.common.exceptions import LockedProfileError, LockingProfileError
+from aiida.manage.profile_access import ProfileAccessManager
 
 ###########################################################################
 # SIMPLE UNIT TESTS
